Create weightplots.py
This code plots the final weights. Supply the weights as a list. when chosen a particular weight, all the weights above that will be highlighted and the neuron numbers are plotted on the plot so that you can verify if these neuron weights are potentiated according to STDP as they are included in the pattern. See pattern.jpeg and weighevolution.jpeg